LPD-68052 Fix imports after moving BasePage to a new folder
