+++ modules/dxp/apps/osb/osb-faro/osb-faro-web/src/main/js/settings/apis/pages/AccessTokenList.tsx
@@ -1,6 +1,6 @@
 import * as API from 'shared/api';
 import Alerts, {AlertTypes} from 'shared/components/Alert';
-import BasePage from 'settings/components/BasePage';
+import BasePage from 'settings/components/base-page/BasePage';
 import Card from 'shared/components/Card';
 import ClayButton from '@clayui/button';
 import ClayLink from '@clayui/link';
