+++ modules/dxp/apps/osb/osb-faro/osb-faro-web/src/main/js/settings/components/data-source/BasePage.tsx
@@ -1,4 +1,4 @@
-import BasePage from 'settings/components/BasePage';
+import BasePage from 'settings/components/base-page/BasePage';
 import DataSourceStatus from './DataSourceStatus';
 import getCN from 'classnames';
 import React from 'react';
