+++ modules/dxp/apps/osb/osb-faro/osb-faro-web/src/main/js/settings/data-privacy/pages/RequestLog.tsx
@@ -1,4 +1,4 @@
-import BasePage from 'settings/components/BasePage';
+import BasePage from 'settings/components/base-page/BasePage';
 import React from 'react';
 import RequestList from '../hocs/RequestList';
 import {getDataPrivacy} from 'shared/util/breadcrumbs';
