+++ modules/dxp/apps/osb/osb-faro/osb-faro-web/src/main/js/settings/definitions/events/pages/BlockList.tsx
@@ -1,4 +1,4 @@
-import BasePage from 'settings/components/BasePage';
+import BasePage from 'settings/components/base-page/BasePage';
 import BlockListCard from '../components/BlockListCard';
 import React from 'react';
 import {getDefinitions} from 'shared/util/breadcrumbs';
