+++ modules/dxp/apps/osb/osb-faro/osb-faro-web/src/main/js/settings/definitions/events/pages/View.tsx
@@ -1,4 +1,4 @@
-import BasePage from 'settings/components/BasePage';
+import BasePage from 'settings/components/base-page/BasePage';
 import EVENT_DEFINITION_QUERY, {
 	EventDefinitionData,
 	EventDefinitionVariables,
