+++ modules/dxp/apps/osb/osb-faro/osb-faro-web/src/main/js/settings/definitions/pages/IndividualAttributes.tsx
@@ -1,5 +1,5 @@
 import * as API from 'shared/api';
-import BasePage from 'settings/components/BasePage';
+import BasePage from 'settings/components/base-page/BasePage';
 import Card from 'shared/components/Card';
 import ClayButton from '@clayui/button';
 import ClayLink from '@clayui/link';
