+++ modules/dxp/apps/osb/osb-faro/osb-faro-web/src/main/js/settings/definitions/pages/Overview.tsx
@@ -1,4 +1,4 @@
-import BasePage from 'settings/components/BasePage';
+import BasePage from 'settings/components/base-page/BasePage';
 import Card from 'shared/components/Card';
 import ClayLink from '@clayui/link';
 import ClayList from '@clayui/list';
