+++ modules/dxp/apps/osb/osb-faro/osb-faro-web/src/main/js/settings/definitions/pages/search/Search.tsx
@@ -1,4 +1,4 @@
-import BasePage from 'settings/components/BasePage';
+import BasePage from 'settings/components/base-page/BasePage';
 import React from 'react';
 import SearchCard from './SearchCard';
 import {getDefinitions} from 'shared/util/breadcrumbs';
