+++ modules/dxp/apps/osb/osb-faro/osb-faro-web/src/main/js/settings/pages/UsageOverviewSaaS.tsx
@@ -1,4 +1,4 @@
-import BasePage from 'settings/components/BasePage';
+import BasePage from 'settings/components/base-page/BasePage';
 
 import ClayLayout from '@clayui/layout';
