+++ modules/dxp/apps/osb/osb-faro/osb-faro-web/src/main/js/settings/pages/Workspace.tsx
@@ -1,6 +1,6 @@
 import * as API from 'shared/api';
 import AddWorkspaceForm from 'shared/components/workspaces/AddWorkspaceForm';
-import BasePage from 'settings/components/BasePage';
+import BasePage from 'settings/components/base-page/BasePage';
 import React from 'react';
 import {addAlert} from 'shared/actions/alerts';
 import {Alert} from 'shared/types';
