+++ modules/dxp/apps/osb/osb-faro/osb-faro-web/src/main/js/settings/pages/data-source/ConfigureCSV.jsx
@@ -1,7 +1,7 @@
 import * as API from 'shared/api';
 import * as breadcrumbs from 'shared/util/breadcrumbs';
 import autobind from 'autobind-decorator';
-import BasePage from 'settings/components/BasePage';
+import BasePage from 'settings/components/base-page/BasePage';
 import DataTransformation, {
 	processFieldMappings
 } from 'settings/components/DataTransformation';
