+++ modules/dxp/apps/osb/osb-faro/osb-faro-web/src/main/js/settings/recommendations/pages/View.tsx
@@ -1,4 +1,4 @@
-import BasePage from 'settings/components/BasePage';
+import BasePage from 'settings/components/base-page/BasePage';
 import Constants, {JobRunStatuses} from 'shared/util/constants';
 import OutputVersionsCard from '../components/OutputVersionsCard';
 import React from 'react';
